fix database FOREIGN KEYS, fix SHOW method for Apartament, fix INDEX methods

+++ database/migrations/2021_08_31_121945_create_cladiri_table.php
@@ -15,7 +15,7 @@ class CreateCladiriTable extends Migration
     {
         Schema::create('cladiri', function (Blueprint $table) {
             $table->id();
-            $table->integer('complex_id');
+            $table->unsignedBigInteger('complex_id');
             $table->string('nume');
             $table->string('numar_etaje');
             $table->softDeletes();
